# packages/@wundr/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wundr API",
    "version": "2.0.0",
    "description": "Comprehensive API for Wundr platform operations including code analysis, batch processing, configuration management, and report generation.",
    "contact": {
      "name": "Wundr Support",
      "url": "https://github.com/adapticai/wundr",
      "email": "support@wundr.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.wundr.io",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/batches": {
      "get": {
        "tags": ["batches"],
        "summary": "Get batch processing jobs",
        "description": "Retrieves batch processing jobs with optional filtering",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Filter batches by type",
            "schema": {
              "type": "string",
              "enum": ["all", "active", "history"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Batches retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BatchJob"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["batches"],
        "summary": "Create new batch job",
        "description": "Creates a new batch processing job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BatchJob"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/batches/{id}": {
      "get": {
        "tags": ["batches"],
        "summary": "Get specific batch job",
        "description": "Retrieves details for a specific batch job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Batch job identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Batch details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BatchJob"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "running", "completed", "failed"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          }
        }
      },
      "CreateBatchRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "configuration": {
            "type": "object"
          }
        },
        "required": ["name", "type"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}